MicrolearnScriptCreator/requirements.txt
streamlit>=1.24.0
google-genai>=1.0.0
fpdf2>=2.5.0
python-dotenv>=1.0.0
pytest>=7.0.0
